escape database contents when rendering
The cleanser helper was not markdown aware and would strip contents from
markdown code blocks. Instead of trying to filter user input before
saving it we save the unmodified input and filter it when rendering with
the help of the :filter_html option in the markdown renderer.

Those implementing tools that directly get their contents from the
database must remember to filter before outputting the contents.