match unmodified URL
When a post's contents are rendered, all URLs are resolved to go around
link shorteners. This has the side effect that short URLs to youtu.be
will be resolved to the full URL, so the embedding renderer cannot rely
on the URL in the `href=` property. Instead, it has to take the
unaltered URL between the <a>..</a> tags.