define available locales
As we are respecting the locale preference of the browser
(HTTP_ACCEPT_LANGUAGE), the browser could request a locale that is not
supported. By defining a list of supported locales, FastGettext will
fall back to en_GB if the requested language is not supported.

Our previous fallback to en_GB would only work if HTTP_ACCEPT_LANGUAGE
was undefined.