save number of comments in data attribute
This fixes 7df/777: "French: Supposedly there is no "more comments"
link"

We used parseInt on a string to get the total number of comments. This
broke for languages other than English. (Actually, `parseInt` returned
NaN in the English UI, which is just as wrong...)

With this commit, the total number of comments is now stored in a data
attribute.