enable markdown syntax extension: media
New syntax for embedding audio files:

~[Audio file title]( http://url-of/first-source.mp3
                     http://url-of/second-source.ogg )

When embedding is enabled, this is converted to:

    <audio controls="true">
      <source src="http://url-of/first-source.mp3 />
      <source src="http://url-of/second-source.ogg />
    </audio>
    <div class="audio-file-title">
      Audio file title
    </div>

The second source is optional.
Mnemonic: "~" looks a bit like a wave form.

~~~~

New syntax for embedding video files:

    %[Video file title]( http://url-of/first-source.mp4
                         http://url-of/second-source.ogv )

When embedding is enabled, this is converted to:

    <video controls="true">
      <source src="http://url-of/first-source.mp4 />
      <source src="http://url-of/second-source.ogv />
    </audio>
    <div class="video-file-title">
      Video file title
    </div>

The second source is optional.
Mnemonic: "%" looks a bit like a position indicator.