home view: add header containing river label
A header with the name of the posts source is displayed for tags and pools
already.  For consistency add a header displaying the label of the currently
selected river on the home view.

This will be hidden on the default themes (see next commit) but may be shown
with other themes.