fix "0 of n comments" bug by refreshing cache
When a comment was added to a cached post that had very recently been
fetched with Post.get_full before the cache expired, and a link to the
comment was followed, post.comments(:from_id => comment.id) would
result in an empty set, as the latest comment would not be part of the
cached set of comments.

The empty set would be rendered as no comments at all, tripping up the
AJAX call to the _comments action in the Comments controller.  The
action would be invoked with an undefined to_id argument.

The fix is to detect that the cached comment set is stale and refresh it
by passing :refresh_cache to the comments method on the cached post.
The fresh set of comments will thus include in latest comment and be
rendered as expected.